# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "仓库管理 API",
    "description": "仓库管理后端接口，包含认证、入库、出库、库存查询、用户信息、配置、图片上传等。业务接口需携带 `Authorization: Bearer <token>`。",
    "version": "1.0.0"
  },
  "paths": {
    "/api/auth/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Register",
        "operationId": "register_api_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/token": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Get Token",
        "operationId": "get_token_api_auth_token_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/inbound": {
      "post": {
        "tags": [
          "inbound"
        ],
        "summary": "Inbound",
        "operationId": "inbound_api_inbound_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InboundRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/outbound": {
      "post": {
        "tags": [
          "outbound"
        ],
        "summary": "Outbound",
        "operationId": "outbound_api_outbound_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OutboundRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/inventory/stats": {
      "get": {
        "tags": [
          "inventory"
        ],
        "summary": "Stats",
        "operationId": "stats_api_inventory_stats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/inventory/overview": {
      "get": {
        "tags": [
          "inventory"
        ],
        "summary": "Overview",
        "operationId": "overview_api_inventory_overview_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/inventory/outbound-list": {
      "get": {
        "tags": [
          "inventory"
        ],
        "summary": "Outbound List",
        "operationId": "outbound_list_api_inventory_outbound_list_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/inventory/statistics-list": {
      "get": {
        "tags": [
          "inventory"
        ],
        "summary": "Statistics List",
        "operationId": "statistics_list_api_inventory_statistics_list_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/user/info": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Get User Info",
        "operationId": "get_user_info_api_user_info_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "user"
        ],
        "summary": "Update User Info",
        "operationId": "update_user_info_api_user_info_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserInfoUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/config": {
      "get": {
        "tags": [
          "config"
        ],
        "summary": "Get Config",
        "operationId": "get_config_api_config_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/upload/image": {
      "post": {
        "tags": [
          "upload"
        ],
        "summary": "Upload Image",
        "operationId": "upload_image_api_upload_image_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_image_api_upload_image_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_upload_image_api_upload_image_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_image_api_upload_image_post"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InboundRequest": {
        "properties": {
          "itemType": {
            "type": "string",
            "title": "Itemtype",
            "description": "物品类型"
          },
          "quantity": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Quantity",
            "description": "数量"
          },
          "expiryDate": {
            "type": "string",
            "title": "Expirydate",
            "description": "过期日期 YYYY-MM-DD"
          },
          "inboundDate": {
            "type": "string",
            "title": "Inbounddate",
            "description": "入库日期 YYYY-MM-DD"
          },
          "productionDate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Productiondate",
            "description": "生产日期",
            "default": ""
          },
          "tag": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tag",
            "description": "标签颜色 hex",
            "default": ""
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location",
            "description": "位置",
            "default": ""
          },
          "photo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo",
            "description": "图片 URL",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "itemType",
          "quantity",
          "expiryDate",
          "inboundDate"
        ],
        "title": "InboundRequest"
      },
      "OutboundRequest": {
        "properties": {
          "itemType": {
            "type": "string",
            "title": "Itemtype",
            "description": "物品类型"
          },
          "quantity": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Quantity",
            "description": "出库数量"
          },
          "outboundDate": {
            "type": "string",
            "title": "Outbounddate",
            "description": "出库日期 YYYY-MM-DD"
          }
        },
        "type": "object",
        "required": [
          "itemType",
          "quantity",
          "outboundDate"
        ],
        "title": "OutboundRequest"
      },
      "RegisterRequest": {
        "properties": {
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "openid": {
            "type": "string",
            "title": "Openid"
          },
          "passcode": {
            "type": "string",
            "title": "Passcode"
          }
        },
        "type": "object",
        "required": [
          "phone",
          "openid",
          "passcode"
        ],
        "title": "RegisterRequest"
      },
      "TokenRequest": {
        "properties": {
          "openid": {
            "type": "string",
            "title": "Openid"
          },
          "passcode": {
            "type": "string",
            "title": "Passcode"
          }
        },
        "type": "object",
        "required": [
          "openid",
          "passcode"
        ],
        "title": "TokenRequest"
      },
      "UserInfoUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "姓名"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone",
            "description": "11位手机号"
          }
        },
        "type": "object",
        "title": "UserInfoUpdate"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}